Move PSI and Search components and update imports.
