+++ assets/js/modules/search-console/components/dashboard/DashboardSearchFunnel.js
@@ -1,5 +1,5 @@
 /**
- * DashboardDetailsSearchFunnel component.
+ * DashboardSearchFunnel component.
  *
  * Site Kit by Google, Copyright 2019 Google LLC
  *
@@ -17,6 +17,11 @@
  */
 
 /**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
  * WordPress dependencies
  */
 import { Component, Fragment } from '@wordpress/element';
@@ -25,14 +30,20 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import Layout from '../../../components/layout/layout';
-import DashboardSearchFunnelInner from '../dashboard/dashboard-widget-search-funnel-inner';
-import { getModulesData } from '../../../util';
-import AnalyticsInactiveCTA from '../../../components/analytics-inactive-cta';
-import DashboardModuleHeader from '../../../components/dashboard/dashboard-module-header';
+import { getModulesData } from '../../../../util';
+import Layout from '../../../../components/layout/layout';
+import DashboardModuleHeader from '../../../../components/dashboard/dashboard-module-header';
+import AnalyticsInactiveCTA from '../../../../components/analytics-inactive-cta';
+import DashboardSearchFunnelInner from './DashboardSearchFunnelInner';
 
-class DashboardDetailsSearchFunnel extends Component {
+class DashboardSearchFunnel extends Component {
 	render() {
+		const modulesData = getModulesData();
+		const { canManageOptions } = global._googlesitekitLegacyData.permissions;
+
+		// Users without manage options capability will not see Setup CTA.
+		const wrapperCols = ! modulesData.analytics.active && ! canManageOptions ? 6 : 12;
+
 		return (
 			<Fragment>
 				<div className="
@@ -40,17 +51,20 @@ class DashboardDetailsSearchFunnel extends Component {
 					mdc-layout-grid__cell--span-12
 				">
 					<DashboardModuleHeader
-						description={ __( 'How your site appeared in Search results and how many visitors you got from Search.', 'google-site-kit' ) }
 						title={ __( 'Search Funnel', 'google-site-kit' ) }
+						description={ __( 'How your site appeared in Search results and how many visitors you got from Search.', 'google-site-kit' ) }
 					/>
 				</div>
-				<div className="mdc-layout-grid__cell mdc-layout-grid__cell--span-12">
+				<div className={ classnames(
+					'mdc-layout-grid__cell',
+					`mdc-layout-grid__cell--span-${ wrapperCols }`
+				) }>
 					<Layout className="googlesitekit-analytics-search-funnel">
 						<div className="mdc-layout-grid">
 							<div className="mdc-layout-grid__inner">
 								<DashboardSearchFunnelInner />
 								{ // Show the Analytics CTA if analytics is not enabled.
-									( ! getModulesData().analytics.active ) &&
+									( ! modulesData.analytics.active ) &&
 									<div className="
 										mdc-layout-grid__cell
 										mdc-layout-grid__cell--span-4-phone
@@ -69,4 +83,4 @@ class DashboardDetailsSearchFunnel extends Component {
 	}
 }
 
-export default DashboardDetailsSearchFunnel;
+export default DashboardSearchFunnel;
