+++ assets/js/modules/search-console/components/dashboard/SearchConsoleDashboardWidgetKeywordTable.js
@@ -30,11 +30,11 @@ import { Component } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { getTimeInSeconds, numberFormat, getModulesData } from '../../../util';
-import withData from '../../../components/higherorder/withdata';
-import { TYPE_MODULES } from '../../../components/data';
-import { getDataTableFromData, TableOverflowContainer } from '../../../components/data-table';
-import PreviewTable from '../../../components/preview-table';
+import { getTimeInSeconds, numberFormat, getModulesData } from '../../../../util';
+import withData from '../../../../components/higherorder/withdata';
+import { TYPE_MODULES } from '../../../../components/data';
+import { getDataTableFromData, TableOverflowContainer } from '../../../../components/data-table';
+import PreviewTable from '../../../../components/preview-table';
 
 class SearchConsoleDashboardWidgetKeywordTable extends Component {
 	render() {
