+++ assets/js/modules/search-console/components/dashboard/SearchConsoleDashboardWidgetOverview.js
@@ -25,11 +25,11 @@ import { Component, Fragment } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { getTimeInSeconds } from '../../../util';
-import DataBlock from '../../../components/data-block';
-import withData from '../../../components/higherorder/withdata';
-import { TYPE_MODULES } from '../../../components/data';
-import PreviewBlock from '../../../components/preview-block';
+import { getTimeInSeconds } from '../../../../util';
+import DataBlock from '../../../../components/data-block';
+import withData from '../../../../components/higherorder/withdata';
+import { TYPE_MODULES } from '../../../../components/data';
+import PreviewBlock from '../../../../components/preview-block';
 import {
 	extractSearchConsoleDashboardData,
 	isDataZeroSearchConsole,
